rename ntt_field to ntt_fields
